Added serverside validation (simple)

+++ TestApplications/Socket.IO_TestApp/Server/service.js
@@ -1,27 +1,33 @@
 var database = require('./database')
 
 function verifyLogIn(username, password){
-    database.verifyLogIn(username, password);
+    if(username && password)
+        database.verifyLogIn(username, password);
 }
 
 function getBidsByUser(userId){
-    database.getBidsByUser(userId);
+    if(userId)
+        database.getBidsByUser(userId);
 }
 
 function placeBid(itemno, userId, value){
-    database.placeBid(itemno, userId, value);
+    if(itemno && userId && value)
+        database.placeBid(itemno, userId, value);
 }
 
 function registerUser(username, firstname, lastname, adress, password){
-    database.registerUser(username, firstname, lastname, adress, password);
+    if(username && firstname && lastname && adress && password)
+        database.registerUser(username, firstname, lastname, adress, password);
 }
 
 function registerItem(name, price, expires, description, addedById){
-    database.registerItem(name, price, expires, description, addedById)
+    if(name && price && expires && description && addedById)
+        database.registerItem(name, price, expires, description, addedById)
 }
 
 function deleteItem(itemno){
-    database.deleteItem(itemno);
+    if(itemno)
+        database.deleteItem(itemno);
 }
 
 function getAllItems(){
@@ -29,11 +35,13 @@ function getAllItems(){
 }
 
 function getLatestBid(bidId){
-    database.getLatestBid(bidId);
+    if(bidId)
+        database.getLatestBid(bidId);
 }
 
 function getLatestItem(itemno){
-    database.getLatestItem((itemno));
+    if(itemno)
+        database.getLatestItem((itemno));
 }
 
 exports.verifyLogIn = verifyLogIn;
